Split two grandfathered over-1000-line files, drop their biome overrides
test/test-utils/db-helpers.ts and test/lib/server-payments-success.test.ts
were both grandfathered into biome.json's noExcessiveLinesPerFile override
list. Split each into topic files under 400 lines and removed their
entries from the override list.

db-helpers.ts becomes test/test-utils/db-helpers/{request,listing-forms,
listings,attendees,attendee-payments,groups,holidays,built-sites,misc}.ts,
re-exported through the existing test/test-utils.ts barrel. Updated the
handful of call sites that imported test-utils/db-helpers.ts directly to
point at the specific submodule.

server-payments-success.test.ts becomes four files split by concern
(basic/package booking, refunds, thank-you/replay, token verification).

Splitting surfaced two jscpd duplicates that the original oversized files
were too large for jscpd to scan (a known side effect per AGENTS.md):
extracted fillSoleCapacityListing (shared with
server-webhooks/multi-ticket-booking.test.ts) and
expectMembersBookedOnDate (shared within the new basic-tests file) to
bring duplication back to 0%.
