+++ test/lib/server-bulk-email/notes-and-history.test.ts
@@ -19,10 +19,8 @@ import {
   getTestPrivateKey,
   useFetchStub,
 } from "#test-utils";
-import {
-  createTestAttendeeDirect,
-  createTestListing,
-} from "#test-utils/db-helpers.ts";
+import { createTestAttendeeDirect } from "#test-utils/db-helpers/attendees.ts";
+import { createTestListing } from "#test-utils/db-helpers/listings.ts";
 import { seedDraft, seedListingWithAttendees, useResend } from "./helpers.ts";
 
 describeWithEnv("server bulk email > notes and history", { db: true }, () => {
