+++ test/test-utils.ts
@@ -5,7 +5,14 @@ export * from "./test-utils/assertions.ts";
 export { getTestDataKey, getTestPrivateKey } from "./test-utils/crypto.ts";
 export * from "./test-utils/csrf.ts";
 export * from "./test-utils/db.ts";
-export * from "./test-utils/db-helpers.ts";
+export * from "./test-utils/db-helpers/attendee-payments.ts";
+export * from "./test-utils/db-helpers/attendees.ts";
+export * from "./test-utils/db-helpers/built-sites.ts";
+export * from "./test-utils/db-helpers/groups.ts";
+export * from "./test-utils/db-helpers/holidays.ts";
+export * from "./test-utils/db-helpers/listing-forms.ts";
+export * from "./test-utils/db-helpers/listings.ts";
+export * from "./test-utils/db-helpers/misc.ts";
 export * from "./test-utils/e2e.ts";
 export * from "./test-utils/email.ts";
 export * from "./test-utils/env.ts";
